extract common code to common package + cleanup
